% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{assert_weights}
\alias{assert_weights}
\title{Assert weights given data set}
\usage{
assert_weights(weights, X)
}
\arguments{
\item{weights}{weights of observations}

\item{X}{data matrix}
}
\value{
weights
}
\description{
Assert weights given data set
}
\examples{
w <- c(1, 2, 3)
X <- matrix(1 : 6, 3, 2)
assert_weights(w, X)

\dontrun{
w <- c(-1, 2, 3)
X <- matrix(1 : 6, 3, 2)
assert_weights(w, X)
}
}
