% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{bic}
\alias{bic}
\title{Compute BIC}
\usage{
bic(model, X, weights = rep(1, nrow(X)), ...)
}
\arguments{
\item{model}{object of class \code{mdl}}

\item{X}{data matrix}

\item{weights}{weights of observations}

\item{...}{arguments to be passed to \code{log_lik}}
}
\value{
BIC, a numeric value
}
\description{
Compute BIC
}
