% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{compute_belongings}
\alias{compute_belongings}
\title{Compute belongings, which is useful in clustering EM algorithm (E-step)}
\usage{
compute_belongings(model, X, ...)
}
\arguments{
\item{model}{object of class \code{mdl}}

\item{X}{data matrix}

\item{...}{arguments passed to \code{log_prob}}
}
\value{
the Z matrix (n-by-g)
}
\description{
Compute belongings, which is useful in clustering EM algorithm (E-step)
}
\examples{
gm1 <- gaussian_mdl(c(1, 2), matrix(c(1, 0, 0, 2), 2, 2))
gm2 <- gaussian_mdl(c(-5, -3), matrix(c(2, -1, -1, 1), 2, 2))
gm3 <- gaussian_mdl(c(0, 0), matrix(c(1, 0.3, 0.3, 1), 2, 2))
gmm <- mixture_mdl(c(0.3, 0.2, 0.5), list(gm1, gm2, gm3))
X <- generate_data(gmm, 1000)
Z <- compute_belongings(gmm, X)
}
