% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{dof}
\alias{dof}
\title{S3 generic \code{dof} to compute degree of freedom}
\usage{
dof(model)
}
\arguments{
\item{model}{object of class \code{mdl}}
}
\value{
degree of freedom, a numeric value
}
\description{
S3 generic \code{dof} to compute degree of freedom
}
