% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{fit}
\alias{fit}
\title{S3 generic \code{fit} to fit model parameter given data set}
\usage{
fit(model, X, weights, ...)
}
\arguments{
\item{model}{object of class \code{mdl}}

\item{X}{data matrix}

\item{weights}{weights of observations}

\item{...}{arguments to be passed to methods}
}
\value{
fitted model
}
\description{
S3 generic \code{fit} to fit model parameter given data set
}
