% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{gaussian_mdl}
\alias{gaussian_mdl}
\title{Create an object of S3 class \code{gaussian_mdl}}
\usage{
gaussian_mdl(mu, sigma, hyper_params = list())
}
\arguments{
\item{mu}{a vector giving the means of the variables}

\item{sigma}{a positive-definite symmetric matrix specifying the covariance matrix of the variables}

\item{hyper_params}{hyperparameters (not implemented)}
}
\value{
object of class \code{gaussian_mdl}
}
\description{
Create an object of S3 class \code{gaussian_mdl}
}
\examples{
gaussian_mdl(c(1, 2), matrix(c(1, 0, 0, 2), 2, 2))
}
