% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{generate_data}
\alias{generate_data}
\title{S3 generic \code{generate_data}}
\usage{
generate_data(n, model, ...)
}
\arguments{
\item{n}{sample size}

\item{model}{object of class \code{mdl}}

\item{...}{}
}
\value{
generated data, usually a matrix with \code{n} rows
}
\description{
S3 generic \code{generate_data}
}
