% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ising.R
\name{gibbs_ising}
\alias{gibbs_ising}
\title{Gibbs Sampler for Generalized Ising Model}
\usage{
gibbs_ising(u, h, J, n_step = 1)
}
\arguments{
\item{u}{a binary ({-1, 1}-valued) vector or matrix. If \code{u} is a matrix, each
row is taken as an input vector.}

\item{h}{a vector giving the external field.}

\item{J}{a symmetric matrix with zero diagonals giving the interaction. If
\code{J} is \code{NULL}, the model degenerates to the case of independent binary.}

\item{n_step}{the number of Gibbs steps, default = 1}
}
\value{
config after Gibbs updates
}
\description{
\code{gibbs_ising} is a Gibbs sampler generates data from Ising model, provided
starting config and model parameters. It is recommended that user uses a
multi-chain approach, that is, providing a matrix \code{u}.
}
\examples{
n <- 1000
p <- 3
h <- rnorm(p)
J <- matrix(0, p, p)
u <- matrix(2 * (runif(n * p) < 0.5) - 1, n, p)
gibbs_ising(u, h, J)
}
