% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ising.R
\name{hamiltonian_ising}
\alias{hamiltonian_ising}
\title{Hamiltonian of Generalized Ising Model}
\usage{
hamiltonian_ising(u, h, J = NULL)
}
\arguments{
\item{u}{a binary ({-1, 1}-valued) vector or matrix. If \code{u} is a matrix, each
row is taken as an input vector.}

\item{h}{a vector giving the external field.}

\item{J}{a symmetric matrix with zero diagonals giving the interaction. If
\code{J} is \code{NULL}, the model degenerates to the case of independent binary.}
}
\value{
a vector giving the energy of configuration u
}
\description{
\code{hamiltonian_ising} is used to compute the energy of given configuration in a
generalized Ising model.
}
\details{
The generalized Ising model (also known as quadratic exponential binary
distribution) is generalizing the adjacent coupling of Ising model to two-way
interactions. The probability mass function of generalized Ising model is
\deqn{P(u \vert h, J) = \frac{1}{Z(h, J)} e^{- H(u; h, J)},}
where \eqn{H(u; h, J) = - h^T u - \frac{1}{2} u^T J u} is the Hamiltonian.
}
\examples{
hamiltonian_ising(
  u = c(+1, +1),
  h = c(0, 0),
  J = matrix(c(0, -1, -1, 0), 2, 2)
)

hamiltonian_ising(
  u = matrix(c(-1, +1, +1, -1), 2, 2, byrow = TRUE),
  h = c(2, 3),
  J = matrix(0, 2, 2)
)
}
