% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{init_belongings}
\alias{init_belongings}
\title{Initialize belongings}
\usage{
init_belongings(n, g)
}
\arguments{
\item{n}{sample size.}

\item{g}{number of clusters.}
}
\value{
a n-by-g belonging matrix.
}
\description{
Initialize belongings
}
\examples{
init_belongings(100, 3)
}
