% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{init_params}
\alias{init_params}
\title{S3 generic \code{init_params} to initialize model parameters given data set}
\usage{
init_params(model, X, weights = rep(1, nrow(X)), ...)
}
\arguments{
\item{model}{object of class \code{mdl}}

\item{X}{data matrix}

\item{weights}{weights of observations}

\item{...}{arguments to be passed to methods}
}
\value{
model with parameters initialized
}
\description{
S3 generic \code{init_params} to initialize model parameters given data set
}
