% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{log_lik}
\alias{log_lik}
\title{Compute Log-likelihood}
\usage{
log_lik(model, X, weights = rep(1, nrow(X)), ...)
}
\arguments{
\item{model}{object of class \code{mdl}}

\item{X}{data matrix}

\item{weights}{weights of observations}

\item{...}{arguments to be passed to \code{log_prob}}
}
\value{
log-likelihood, a numeric value
}
\description{
Compute Log-likelihood
}
