% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{log_prob}
\alias{log_prob}
\title{S3 generic \code{log_prob} to compute log probability for each data point}
\usage{
log_prob(model, X, ...)
}
\arguments{
\item{model}{object of class \code{mdl}}

\item{X}{data matrix}

\item{...}{arguments to be passed to methods}
}
\value{
log probability, a numeric vector
}
\description{
S3 generic \code{log_prob} to compute log probability for each data point
}
