% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log-sum-exp}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{a numeric object}
}
\value{
\code{log(sum(exp(x)))}
}
\description{
Log-sum-exp
}
\examples{
log_sum_exp(c(1, 2, 30))
}
