% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logistic}
\alias{logistic}
\title{Standard Logistic Function}
\usage{
logistic(x)
}
\arguments{
\item{x}{a numeric input}
}
\value{
a (0, 1)-ranged output of the same size as \code{x}
}
\description{
Equivalent to \code{\link[stats:Logistic]{stats::plogis()}} with default arguments.
}
\details{
The standard logistic function is defined as
\deqn{p(x) = \frac{1}{1 + e^{-x}}.}
We decided to write our own implementation instead of using the C/C++ powered
\code{\link[stats:Logistic]{stats::plogis()}} after benchmarking performance.
}
\examples{
logistic(0)
}
