% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{new_gaussian_mdl}
\alias{new_gaussian_mdl}
\title{Constructor of S3 class \code{gaussian_mdl}, it is a subclass of \code{mdl}}
\usage{
new_gaussian_mdl(params, hyper_params)
}
\arguments{
\item{hyper_params}{}
}
\value{
object of class \code{gaussian_mdl}
}
\description{
Constructor of S3 class \code{gaussian_mdl}, it is a subclass of \code{mdl}
}
