% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{new_mdl}
\alias{new_mdl}
\title{Constructor of S3 class \code{mdl}}
\usage{
new_mdl(params, hyper_params, ..., class = character())
}
\arguments{
\item{params}{parameters of the model}

\item{hyper_params}{hyper parameters (e.g. learning rate, regularization)}

\item{...}{allowing subclass}

\item{class}{allowing subclass}
}
\value{
newly created object
}
\description{
Constructor of S3 class \code{mdl}
}
