% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{new_mixture_mdl}
\alias{new_mixture_mdl}
\title{Constructor of S3 class \code{mixture_mdl}, it is a subclass of \code{mdl}}
\usage{
new_mixture_mdl(params, hyper_params)
}
\arguments{
\item{params}{parameters of the model}

\item{hyper_params}{hyper parameters (e.g. learning rate, regularization)}
}
\value{
object of class \code{mixture_mdl}
}
\description{
Constructor of S3 class \code{mixture_mdl}, it is a subclass of \code{mdl}
}
