% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdirichlet}
\alias{rdirichlet}
\title{RNG for Dirichlet Distribution}
\usage{
rdirichlet(n, alphas)
}
\arguments{
\item{n}{number of observations.}

\item{alphas}{vector of concentration parameters.}
}
\value{
a matrix where each row is one observation.
}
\description{
RNG for Dirichlet Distribution
}
\examples{
Z <- rdirichlet(10, c(2, 4, 8))
}
