% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowwise_kronecker}
\alias{rowwise_kronecker}
\title{RowWise Kronecker}
\usage{
rowwise_kronecker(x, y)
}
\arguments{
\item{x}{the first data matrix}

\item{y}{the second data matrix, should have the same number of rows as \code{x}}
}
\value{
transformed matrix
}
\description{
\code{rowwise_kronecker} is a rowwise Kronecker utility function. It is designed
for feature engineering, but we use it to create quadratic/interaction terms
of config.
}
\examples{
x <- matrix(
  c(
    1, 1,
    1, 2,
    1, 3
  ),
  3, 2, byrow = TRUE
)
y <- matrix(
  c(
    1, 0,
    0, 1,
    1, 1
  ),
  3, 2, byrow = TRUE
)
rowwise_kronecker(x, y)
}
