% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleep}
\alias{sleep}
\title{Dataset: sleep}
\format{
A data frame with 1015 rows (each is a sleep event) and 4 variables:
\describe{
\item{subject_id}{subject id}
\item{date}{date of sleep onset}
\item{sleep_onset}{sleep time}
\item{wake_onset}{awake time}
}
}
\source{
Ying Chen, \email{Ying.Chen@nyspi.columbia.edu}
}
\usage{
sleep
}
\description{
A dataset containing sleep/awake onsets of sleep events.
}
\keyword{datasets}
