% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{soft_max}
\alias{soft_max}
\title{Softmax}
\usage{
soft_max(x)
}
\arguments{
\item{x}{a numeric object}
}
\value{
softmax, define as \code{exp(x) / sum(exp(x))}
}
\description{
Softmax
}
\examples{
soft_max(c(1, 2, 30))
}
