% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{validate_gaussian_mdl}
\alias{validate_gaussian_mdl}
\title{Validator of S3 class \code{gaussian_mdl}, it is a subclass of \code{mdl}}
\usage{
validate_gaussian_mdl(model)
}
\arguments{
\item{model}{object of class \code{gaussian_mdl}}
}
\value{
itself
}
\description{
Validator of S3 class \code{gaussian_mdl}, it is a subclass of \code{mdl}
}
