% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{validate_mixture_mdl}
\alias{validate_mixture_mdl}
\title{Validator of S3 class \code{mixture_mdl}, it is a subclass of \code{mdl}}
\usage{
validate_mixture_mdl(model)
}
\arguments{
\item{model}{object of class \code{mixture_mdl}}
}
\value{
itself
}
\description{
Validator of S3 class \code{mixture_mdl}, it is a subclass of \code{mdl}
}
